refactor(agentos_router): extract _find_valid_tier into shared tiers_util
Move _find_valid_tier out of v4_phase3.py into a new tiers_util module so
V4Phase3Strategy and the incoming PilotStrategy share one implementation.
Behavior-preserving: v4_phase3 now imports it; TIER_ORDER walk and fallback
semantics are unchanged.
